Turning off lights tea 88 (#10)
* Switched Baked Emissions to Point Lights

Switched from using baked emission lighting to point lights + Added the lights as a prefab so they can be changed all at once + Increased the amount of lights able to affect an object from 4 to 8

* Trying to get Dictionaries to Display in Inspector

Dictionaries do not display by default. I am trying to get them to so I can have a dictionary containing a mesh and its corresponding light

* Added Serialized Dictionary package

Added in the Serialized Dictionary package by 'ayellowpaper' to replace the custom one I was working on.

* Swapped over to Odin Inspector system

Swapped over to the Odin Inspector system for Serializable Objects to use for Dictionaries. It was included by another team member but forgotten about, but has a much simpler system for serializing objects than others.

* Added Dictionary Objects + Light Switch Awake Method

Added Dictionary Objects which contain a pair of light objects and meshes which the user can set + The Light Switch method checks and sets up the lights in the dictionary based on if they are set to on or off with the lightOnOrOff boolean

* Raycasting & Outline Working

Raycasting works to identify the object with the LightSwitch tag and then turn on the outline for it

* Update LightSwitch.cs

Added a check to see if an outline has been added to the switch, and if not, to add one

* Trying to get lights to turn on and off

Can get the lightOnOrOff variable to change which means it is swapping states, but cannot get the lights to change after startup

* Got Light Switches Working (Thanks to James)

Was able to get the light switches working finally thanks to help from James (WOLFLEADER)

* updates light switch

* Trying to fix the outlines and light switches

* fixed minor bug

* Things

* Set Outline Parameters + Connected Remaining Lights

Set the outline parameters (Colour, width, mode) + Connected the remaining lights to their switches

* Lights & Outlines fixed

* Fixed some broken materials in the test scene

---------

Co-authored-by: Wolfleader <james.morgan@iinet.net.au>