Street 10 houses tea 120 (#17)
* Made Adjustments to House + File Cleanup

Made some adjustments to the house model to fix some areas I wasn't happy with. Currently the ground is missing as I am trying to figure out how to transition over to using the Terrain system, but having some trouble figuring out how to get the material to not clip through the ground

Cleaned up some of the package folders and files and placed them into a "Packages" folder so the root directory wasn't as cluttered

* Added Australian Asset Pack + Road Generator

I have moved all of the packages we have downloaded into a 'Packages' folder so it does not clog up the root directory

Added a gitignore for Packages/Assets/ so that they don't go to the repo. Some can be quite large, such as the Australian Flora assets, so they need to be redownloaded instead

* Fixed Shader Errors + Added to gitignore

Fixed errors within the DEC shaders that was causing my Australian Flora assets to have pink materials

Added the Australian Flora assets to gitignore specifically in case it gets reimported, as it will import to the assets folder

* Finished Tutorial House + Implementing Assets

Finally finished the details for the tutorial house, as well as made a second slightly altered house model, as well as accompanying dummy houses (Empty houses) to use to fill the street

Finished implementing the assets from the Australian Flora pack

* Finishing Up

Creating final prefabs and placing objects

* Finished for now

All 10 houses are in there, and prefabs have been made for them (We can use the prefabs to edit them all at once instead of needing to go one at a time)

* Added game_Scene Scene file

Added the game_Scene file as a sort of final file that we can place the culmination of our work into. Should only be worked on one at a time to not cause merge conflicts

* Reorganized file structure

Reorganized the file structure of the project as there were files all over the place