Performance adjustment (#20)
* Adjusted LOD + Enabled Occlusion Culling

Adjusted the LOD on the trees so they transition into lower-poly models sooner

Enabled Occlusion Culling and baked it. Now will cull anything the camera cannot see

* Added Prop Pack + Finished One House Prefab

Added the House Furniture Pack by Finward Studios (Avalible for download on the shared folder)

Finished a test prefab for the houses to see what the props will look like in the scene

* Update OutdoorScene.unity

* Created House Prefab Variations

Created a couple of variations on the house prefabs to give them more variety. There are a couple of different house configurations now with different prop placements

* Update OutdoorScene.unity